Update issue 92 and issue 61
Updated the support for executing a servlet engine (Jetty) together with a mule-junit test or standalone to Mule 3 new property api, providing the mule servlet with mule-context.

ISSUE: MuleServerWithServletContainer still is based in a THread-sleep() to sync the startup, should use a listener on the muleContext to know when it is started and then initiare the servlet container.

ISSUE: service-project's pom.xml file now have an explicit dependency to jetty in test scope. Should be addressed by issue 61!